rename to `curve`
